requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Task Queue
celery==5.3.4

# AI/ML
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
pinecone-client==3.0.0
sentence-transformers==2.2.2

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pydantic==2.5.2

# Monitoring & Logging
sentry-sdk[flask]==1.39.1

# Export
reportlab==4.0.7  # For PDF generation
cairosvg==2.7.1   # For SVG to PNG conversion

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0